+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+* Added `ExtraBodyParameters` option to text and document translation methods to pass arbitrary parameters in the request body. This can be used to access beta features or override built-in parameters (such as `target_lang`, `source_lang`, etc.).
+
 ## [1.15.0] - 2025-04-25
 ### Added
 * Added support for the /v3 Multilingual Glossary APIs in the client library 
@@ -200,6 +205,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2021-11-05
 Initial release.
 
+[Unreleased]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.15.0...HEAD
 [1.15.0]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.14.0...v1.15.0
 [1.14.0]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.13.0...v1.14.0
 [1.13.0]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.12.0...v1.13.0